# BookingMovieTickets/Frontend/public/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BookingMovieTickets",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/json;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;IEEE754Compatible=false": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/json;IEEE754Compatible=true": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authentication for Booking Movie Tickets",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}